Migrate some installer-specific Disks methods to InstallerDiskOps trait
