+++ Cargo.toml
@@ -64,6 +64,7 @@ proc-mounts = "0.1.0"
 rayon = "1.0.2"
 sys-mount = "1.1.0"
 tempdir = "0.3.7"
+bitflags = "1.0.4"
 
 [dependencies.failure]
 version = "0.1.2"
