+++ crates/disks/src/config/disks.rs
@@ -1,31 +1,29 @@
-use FileSystem::*;
 use disk_types::{BlockDeviceExt, PartitionExt, PartitionTableExt, SectorExt};
 use itertools::Itertools;
 use libparted::{Device, DeviceType};
-use misc::{self, hasher};
+use misc;
 use proc_mounts::{MOUNTS, SWAPS};
 use external::{
-    cryptsetup_close, cryptsetup_open, generate_unique_id, lvs,
+    cryptsetup_close, cryptsetup_open, lvs,
     physical_volumes_to_deactivate, pvs, vgdeactivate, CloseBy
 };
 use partition_identity::PartitionID;
 use rayon::iter::IntoParallelRefIterator;
 use rayon::prelude::*;
-use std::{io, str, thread};
-use std::borrow::Cow;
-use std::collections::HashSet;
-use std::ffi::OsStr;
+use std::{fs, io, str, thread};
+use std::collections::{BTreeMap, HashSet};
 use std::ffi::OsString;
 use std::iter::{self, FromIterator};
+use std::os::unix::ffi::{OsStrExt, OsStringExt};
 use std::path::{Path, PathBuf};
 use std::time::Duration;
 use super::{detect_fs_on_device, find_partition, find_partition_mut, Disk, LvmEncryption, PartitionTable, PVS};
 use super::partitions::{FORMAT, REMOVE, SOURCE};
 use super::super::{
-    Bootloader, DecryptionError, DiskError, DiskExt, FileSystem, FileSystemSupport,
+    Bootloader, DecryptionError, DiskError, DiskExt, FileSystem,
     PartitionFlag, PartitionInfo, LogicalDevice
 };
-use sys_mount::{unmount, swapoff, UnmountFlags};
+use sys_mount::{unmount, swapoff, Mount, MountFlags, Mounts, Unmount, UnmountFlags};
 
 /// A configuration of disks, both physical and logical.
 #[derive(Debug, Default, PartialEq)]
@@ -171,6 +169,78 @@ impl Disks {
         Box::new(iterator)
     }
 
+    /// Mounts all targets in this disks object.
+    pub fn mount_all_targets<P: AsRef<Path>>(
+        &self,
+        base_dir: P
+    ) -> io::Result<Mounts> {
+        let base_dir = base_dir.as_ref();
+        let targets = self.get_partitions()
+            .filter(|part| part.target.is_some() && part.filesystem.is_some());
+
+        // The mount path will actually consist of the target concatenated with the
+        // root. NOTE: It is assumed that the target is an absolute path.
+        let paths: BTreeMap<PathBuf, (PathBuf, &'static str)> = targets
+            .map(|target| {
+                // Path mangling commences here, since we need to concatenate an absolute
+                // path onto another absolute path, and the standard library opts for
+                // overwriting the original path when doing that.
+                let target_mount: PathBuf = {
+                    // Ensure that the base_dir path has the ending '/'.
+                    let base_dir = base_dir.as_os_str().as_bytes();
+                    let mut target_mount: Vec<u8> = if base_dir[base_dir.len() - 1] == b'/' {
+                        base_dir.to_owned()
+                    } else {
+                        let mut temp = base_dir.to_owned();
+                        temp.push(b'/');
+                        temp
+                    };
+
+                    // Cut the starting '/' from the target path if it exists.
+                    let target_path = target.target.as_ref().unwrap().as_os_str().as_bytes();
+                    let target_path = if ! target_path.is_empty() && target_path[0] == b'/' {
+                        if target_path.len() > 1 { &target_path[1..] } else { b"" }
+                    } else {
+                        target_path
+                    };
+
+                    // Append the target path to the base_dir, and return it as a path type.
+                    target_mount.extend_from_slice(target_path);
+                    PathBuf::from(OsString::from_vec(target_mount))
+                };
+
+                let fs = match target.filesystem.unwrap() {
+                    FileSystem::Fat16 | FileSystem::Fat32 => "vfat",
+                    fs => fs.into(),
+                };
+
+                (target_mount, (target.device_path.clone(), fs))
+            })
+            .collect();
+
+        // Each mount directory will be created and then mounted before progressing to
+        // the next mount in the map. The BTreeMap that the mount targets were
+        // collected into will ensure that mounts are created and mounted in
+        // the correct order.
+        let mut mounts = Vec::new();
+
+        for (target_mount, (device_path, filesystem)) in paths {
+            if let Err(why) = fs::create_dir_all(&target_mount) {
+                error!("unable to create '{}': {}", why, target_mount.display());
+            }
+
+            mounts.push(Mount::new(
+                device_path,
+                &target_mount,
+                filesystem,
+                MountFlags::empty(),
+                None,
+            )?.into_unmount_drop(UnmountFlags::DETACH));
+        }
+
+        Ok(Mounts(mounts))
+    }
+
     /// Get all partitions across all physical and logical devices.
     pub fn get_partitions<'a>(&'a self) -> Box<Iterator<Item = &'a PartitionInfo> + 'a> {
         Box::new(self.get_physical_partitions().chain(self.get_logical_partitions()))
@@ -255,27 +325,6 @@ impl Disks {
         PartitionID::new_uuid(target)
             .get_device_path()
             .and_then(move |ref target| self.get_partition_by_path_mut(target))
-    }
-
-    /// Reports file systems that need to be supported in the install.
-    pub fn get_support_flags(&self) -> FileSystemSupport {
-        let mut flags = FileSystemSupport::empty();
-
-        for partition in self.get_partitions() {
-            match partition.filesystem {
-                Some(Btrfs) => flags |= FileSystemSupport::BTRFS,
-                Some(Ext2) | Some(Ext3) | Some(Ext4) => flags |= FileSystemSupport::EXT4,
-                Some(F2fs) => flags |= FileSystemSupport::F2FS,
-                Some(Fat16) | Some(Fat32) => flags |= FileSystemSupport::FAT,
-                Some(Ntfs) => flags |= FileSystemSupport::NTFS,
-                Some(Xfs) => flags |= FileSystemSupport::XFS,
-                Some(Luks) => flags |= FileSystemSupport::LUKS,
-                Some(Lvm) => flags |= FileSystemSupport::LVM,
-                _ => continue
-            };
-        }
-
-        flags
     }
 
     /// Deactivates all device maps associated with the inner disks/partitions
@@ -832,119 +881,6 @@ impl Disks {
         }
 
         Ok(())
-    }
-
-    /// Generates the crypttab and fstab files in memory.
-    pub fn generate_fstabs(&self) -> (OsString, OsString) {
-        info!("generating /etc/crypttab & /etc/fstab in memory");
-        let mut crypttab = OsString::with_capacity(1024);
-        let mut fstab = OsString::with_capacity(1024);
-
-        let partitions = self.physical
-            .iter()
-            .flat_map(|x| {
-                x.file_system.as_ref().into_iter()
-                    .chain(x.partitions.iter())
-                    .map(|p| (true, &None, p))
-            })
-            .chain(self.logical.iter().flat_map(|x| {
-                let luks_parent = &x.luks_parent;
-                let is_unencrypted: bool = x.encryption.is_none();
-                x.file_system.as_ref().into_iter()
-                    .chain(x.partitions.iter())
-                    .map(move |p| (is_unencrypted, luks_parent, p))
-            }));
-
-        fn write_fstab(fstab: &mut OsString, partition: &PartitionInfo) {
-            if let Some(entry) = partition.get_block_info() {
-                entry.write_entry(fstab);
-            }
-        }
-
-        let mut swap_uuids: Vec<u64> = Vec::new();
-
-        for (is_unencrypted, luks_parent, partition) in partitions {
-            if let Some(&(_, Some(ref enc))) = partition.volume_group.as_ref() {
-                let password: Cow<'static, OsStr> =
-                    match (enc.password.is_some(), enc.keydata.as_ref()) {
-                        (true, None) => Cow::Borrowed(OsStr::new("none")),
-                        (false, None) => Cow::Borrowed(OsStr::new("/dev/urandom")),
-                        (true, Some(_key)) => unimplemented!(),
-                        (false, Some(&(_, ref key))) => {
-                            let path = key.clone()
-                                .expect("should have been populated")
-                                .1
-                                .join(&enc.physical_volume);
-                            Cow::Owned(path.into_os_string())
-                        }
-                    };
-
-                let path = luks_parent.as_ref().map_or(partition.get_device_path(), |x| &x);
-
-                match PartitionID::get_uuid(path) {
-                    Some(uuid) => {
-                        crypttab.push(&enc.physical_volume);
-                        crypttab.push(" UUID=");
-                        crypttab.push(&uuid.id);
-                        crypttab.push(" ");
-                        crypttab.push(&password);
-                        crypttab.push(" luks\n");
-                        write_fstab(&mut fstab, &partition);
-                    }
-                    None => error!(
-                        "unable to find UUID for {} -- skipping",
-                        partition.get_device_path().display()
-                    ),
-                }
-            } else if partition.is_swap() {
-                if is_unencrypted {
-                    match PartitionID::get_uuid(&partition.get_device_path()) {
-                        Some(uuid) => {
-                            let unique_id = generate_unique_id("cryptswap", &swap_uuids)
-                                .unwrap_or_else(|_| "cryptswap".into());
-
-                            swap_uuids.push(hasher(&unique_id));
-
-                            crypttab.push(&unique_id);
-                            crypttab.push(" UUID=");
-                            crypttab.push(&uuid.id);
-                            crypttab.push(
-                                " /dev/urandom swap,offset=1024,cipher=aes-xts-plain64,size=512\n",
-                            );
-
-                            fstab.push(&[
-                                "/dev/mapper/",
-                                &unique_id,
-                                "  none  swap  defaults  0  0\n",
-                            ].concat());
-                        }
-                        None => error!(
-                            "unable to find UUID for {} -- skipping",
-                            partition.get_device_path().display()
-                        ),
-                    }
-                } else {
-                    fstab.push(partition.get_device_path());
-                    fstab.push("  none  swap  defaults  0  0\n");
-                }
-            } else {
-                write_fstab(&mut fstab, &partition);
-            }
-        }
-
-        info!(
-            "generated the following crypttab data:\n{}",
-            crypttab.to_string_lossy(),
-        );
-
-        info!(
-            "generated the following fstab data:\n{}",
-            fstab.to_string_lossy()
-        );
-
-        crypttab.shrink_to_fit();
-        fstab.shrink_to_fit();
-        (crypttab, fstab)
     }
 
     /// Loads existing logical volume data into memory, excluding encrypted volumes.
