+++ crates/disks/src/config/mod.rs
@@ -15,7 +15,6 @@ pub use self::lvm::*;
 pub use self::partitions::*;
 pub use disk_types::Sector;
 
-use partition_identity::PartitionIdentifiers;
 use std::collections::BTreeMap;
 use std::io;
 use std::path::{Path, PathBuf};
@@ -38,6 +37,7 @@ pub fn get_size(path: &Path) -> io::Result<u64> {
 mod tests {
     use super::*;
     use operations::*;
+    use partition_identity::PartitionIdentifiers;
 
     fn get_default() -> Disks {
         Disks {
