+++ src/installer/mod.rs
@@ -1,3 +1,6 @@
+pub mod bitflags;
+pub mod traits;
+
 mod state;
 mod steps;
 
@@ -13,7 +16,6 @@ use self::state::InstallerState;
 use std::io;
 use std::path::{Path, PathBuf};
 use std::sync::atomic::Ordering;
-use sys_mount::Mounts;
 use tempdir::TempDir;
 use timezones::Region;
 
@@ -153,7 +155,7 @@ impl Installer {
             info!("mounting temporary chroot directory at {}", Self::CHROOT_ROOT);
 
             let mount_dir = TempDir::new(Self::CHROOT_ROOT)?;
-            let mut mounts = Installer::mount(&disks, mount_dir.path())?;
+            let mut mounts = disks.mount_all_targets(mount_dir.path())?;
 
             if PARTITIONING_TEST.load(Ordering::SeqCst) {
                 info!("PARTITION_TEST enabled: exiting before unsquashing");
@@ -374,12 +376,6 @@ impl Installer {
     /// configuration specified.
     fn partition<F: FnMut(i32)>(disks: &mut Disks, callback: F) -> io::Result<()> {
         steps::partition(disks, callback)
-    }
-
-    /// Mount all target paths defined within the provided `disks`
-    /// configuration.
-    fn mount(disks: &Disks, chroot: &Path) -> io::Result<Mounts> {
-        steps::mount(disks, chroot)
     }
 
     /// Extracts the squashfs image into the new install, and then gets the os-release data.
