+++ src/lib.rs
@@ -20,6 +20,8 @@ pub extern crate proc_mounts;
 pub extern crate sys_mount;
 
 #[macro_use]
+extern crate bitflags;
+#[macro_use]
 extern crate cascade;
 extern crate dirs;
 extern crate distinst_utils as misc;
